requirements.txt
torch
torchvision
timm
opencv-python
pydicom
pylibjpeg
pylibjpeg-libjpeg
pylibjpeg-openjpeg
numpy
matplotlib
seaborn
scikit-learn
tqdm
grad-cam
Pillow